Put the actual extension manifest in a separate file
